implemented osc without routing

+++ src/controller/network/OTAController.cpp
@@ -4,7 +4,7 @@
 
 #include "OTAController.h"
 
-OTAController::OTAController(char *deviceName, char *password, uint8_t port) {
+OTAController::OTAController(char *deviceName, char *password, uint16_t port) {
     this->deviceName = deviceName;
     this->password = password;
     this->port = port;
